replace logic to the custom hook, add logic and styles to accessories and tablets catalogs
