--- src/pages/catalogPages/phonesCatalog.module.scss
@@ -1,9 +1,0 @@
-// @import './../../../styles';
-
-// .phone-catalog {
-//   display: grid;
-//   grid-template-columns: repeat(4, 1fr);
-//   row-gap: 40px;
-//   column-gap: 16px;
-//   margin-bottom: 40px;
-// }
